+++ web/packages/new/photos/components/PeopleList.tsx
@@ -1,6 +1,6 @@
 import { useIsMobileWidth } from "@/base/hooks";
 import { faceCrop, unidentifiedFaceIDs } from "@/new/photos/services/ml";
-import type { Person } from "@/new/photos/services/ml/cgroups";
+import type { Person } from "@/new/photos/services/ml/people";
 import type { EnteFile } from "@/new/photos/types/file";
 import { Skeleton, Typography, styled } from "@mui/material";
 import { t } from "i18next";
