+++ web/packages/new/photos/services/ml/index.ts
@@ -18,9 +18,9 @@ import { isInternalUser } from "../feature-flags";
 import { getRemoteFlag, updateRemoteFlag } from "../remote-store";
 import { setSearchPeople } from "../search";
 import type { UploadItem } from "../upload/types";
-import { namedPeopleFromCGroups, syncCGroups, type Person } from "./cgroups";
 import { regenerateFaceCrops } from "./crop";
 import { clearMLDB, getFaceIndex, getIndexableAndIndexedCounts } from "./db";
+import { namedPeopleFromCGroups, syncCGroups, type Person } from "./people";
 import { MLWorker } from "./worker";
 import type { CLIPMatches } from "./worker-types";
