+++ web/packages/new/photos/services/search/index.ts
@@ -4,7 +4,7 @@ import { ComlinkWorker } from "@/base/worker/comlink-worker";
 import { FileType } from "@/media/file-type";
 import i18n, { t } from "i18next";
 import { clipMatches, isMLEnabled, isMLSupported } from "../ml";
-import type { NamedPerson } from "../ml/cgroups";
+import type { NamedPerson } from "../ml/people";
 import type {
     LabelledFileType,
     LabelledSearchDateComponents,
