+++ web/packages/new/photos/services/user-entity.ts
@@ -11,8 +11,8 @@ import { ensure } from "@/utils/ensure";
 import { nullToUndefined } from "@/utils/transform";
 import { z } from "zod";
 import { gunzip } from "./gzip";
-import type { CGroup } from "./ml/cgroups";
 import { applyCGroupDiff } from "./ml/db";
+import type { CGroup } from "./ml/people";
 
 /**
  * User entities are predefined lists of otherwise arbitrary data that the user
